+++ src/django/users/models.py
@@ -54,6 +54,7 @@ class PlanItLocation(models.Model):
     admin = models.CharField(max_length=16, null=False, blank=True)
     api_city_id = models.IntegerField(null=True, blank=True)
     point = models.PointField(srid=4326, null=True, blank=True)
+    georegion = models.ForeignKey(GeoRegion, null=True, blank=True)
     is_coastal = models.BooleanField(default=False)
 
     objects = PlanItLocationManager()