+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add database field to track PlanItLocation georegion directly
 - Prevent Missy ingest script from failing if user organization has same name as Missy organization
 - Prevent Chrome's insubordinate autocomplete on the city selector
+- Add automatic emails notifying users of their upcoming trial expiration
 
 ## [1.0.2] - 2018-04-19
 ### Fixed